+++ fixtures/multi-entry-library/src/index.ts
@@ -2,5 +2,3 @@ export const add = (a: number, b: number) => a + b;
 
 // forces JobSummary into a chunk
 export * from './components';
-
-export {};
