+++ package.json
@@ -21,7 +21,7 @@
     "test:playwright": "playwright test",
     "fixtures:add": "tsx ./scripts/add-fixture.ts",
     "fixtures:dev": "pnpm %fixtures dev",
-    "sync:dependencies": "pnpm --recursive sync-dependencies && pnpm install",
+    "sync:dependencies": "pnpm --recursive sync:dependencies && pnpm install",
     "sync:ignore": "ignore-sync .",
     "prepare-publish": "pnpm bootstrap && pnpm build",
     "release": "pnpm prepare-publish && changeset publish",
@@ -39,7 +39,7 @@
     "@crackle/core": "workspace:*",
     "@playwright/test": "^1.32.0",
     "@preconstruct/eslint-plugin-format-js-tag": "^0.2.0",
-    "@types/fs-extra": "^9.0.13",
+    "@types/fs-extra": "^11.0.1",
     "@vitejs/plugin-react": "^3.0.0",
     "enquirer": "^2.3.6",
     "eslint": "^8.25.0",
