+++ packages/core/package.json
@@ -130,7 +130,7 @@
   "devDependencies": {
     "@types/babel__core": "^7.1.20",
     "@types/express": "^4.17.14",
-    "@types/fs-extra": "^9.0.13",
+    "@types/fs-extra": "^11.0.1",
     "@types/glob-to-regexp": "^0.4.1",
     "@types/node": "^18.11.17",
     "@types/react": "^18.0.25",
