+++ packages/core/src/utils/resolve-from.ts
@@ -5,9 +5,9 @@ import resolveSync from 'resolve-from';
 export const resolveFrom = async (fromDirectory: string, moduleId: string) =>
   resolveSync(fromDirectory, moduleId);
 
+let crackleRoot: string;
+
 export const resolveFromCrackle = (...paths: string[]) => {
-  const crackleRoot = path.dirname(
-    resolveSync('.', '@crackle/core/package.json'),
-  );
+  crackleRoot ||= path.dirname(resolveSync('.', '@crackle/core/package.json'));
   return path.join(crackleRoot, ...paths);
 };
